requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0

# Google Cloud & AI
google-cloud-aiplatform>=1.87.0
google-cloud-core==2.4.1
google-adk==0.1.0
google-generativeai==0.3.2

# Database
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0

